+++ pypeit/images/calibrationimage.py
@@ -1,4 +1,4 @@
-""" Object to hold + process a single image"""
+""" Uber object for calibration images, e.g. arc, flat """
 
 import inspect
 
@@ -7,22 +7,15 @@ import numpy as np
 
 
 from pypeit import msgs
-
-from pypeit.core import combine
 from pypeit.par import pypeitpar
-
-from pypeit.images import pypeitimage
-from pypeit.images import processrawimage
-from pypeit.images import maskimage
-
-
+from pypeit.images import buildimage
 
 from IPython import embed
 
 
-class CalibrationImage(pypeitimage.PypeItImage, maskimage.ImageMask):
+class CalibrationImage(object):
     """
-    Class to generate a combined calibration image from a list of input images or
+    Class to generate (and hold) a combined calibration image from a list of input images or
     simply to hold a previously generated calibration image.
 
     Args:
@@ -30,42 +23,42 @@ class CalibrationImage(pypeitimage.PypeItImage, maskimage.ImageMask):
             Spectrograph used to take the data.
         det (:obj:`int`, optional):
             The 1-indexed detector number to process.
-        proc_par (:class:`pypeit.par.pypeitpar.ProcessImagesPar`):
+        proc_par (:class:`pypeit.par.pypeitpar.ProcessImagesPar` or None):
             Parameters that dictate the processing of the images.  See
             :class:`pypeit.par.pypeitpar.ProcessImagesPar` for the
             defaults.
 
         files (list, optional):
             List of filenames to be combined
-        frametype (str, optional): Frame type
 
     Attributes:
-        image (np.ndarray):
-        file_list (list): List of files to process
-        process_steps (list): List of processing steps to be used
+        pypeitImage (:class:`pypeit.images.pypeitimage.PypeItImage`):
+        file_list (list):
+            List of files to process
+        process_steps (list):
+            List of processing steps to be used
 
     """
     def __init__(self, spectrograph, det, proc_par, files=None):
 
-        # Init me
-        pypeitimage.PypeItImage.__init__(self, spectrograph, det)
+        # Required items
+        self.spectrograph = spectrograph
+        self.det = det
 
         # Assign the internal list of files
         self._set_files(files)
 
         # Required parameters
-        if not isinstance(proc_par, pypeitpar.ProcessImagesPar):
-            msgs.error('Provided ParSet for must be type ProcessImagesPar.')
-        self.proc_par = proc_par  # This musts be named this way as it is frequently a child
-
-        # Internal images
-        self.image = None
-        #self.ivar = None
-        # TODO add here a noise model and put the images in ADU!
-        self.crmask = None
+        if proc_par is not None:
+            if not isinstance(proc_par, pypeitpar.ProcessImagesPar):
+                msgs.error('Provided ParSet for must be type ProcessImagesPar.')
+        self.proc_par = proc_par  # This must be named this way as it is frequently a child
 
         # Process steps
         self.process_steps = []
+
+        # Standard output
+        self.pypeitImage = None
 
     @property
     def nfiles(self):
@@ -125,45 +118,13 @@ class CalibrationImage(pypeitimage.PypeItImage, maskimage.ImageMask):
                 Bad pixel mask
 
         Returns:
-            np.ndarray: Copy of self.image
+            PypeItImage:
+
         """
-        # Check there are files to combine
-        if self.nfiles == 0:
-            msgs.error('No files to combine!')
-
-        # Load up image array
-        shape = None
-        image_arr = None
-        self.binning = None
-        for kk,file in enumerate(self.file_list):
-            # Process raw file
-            processrawImage = processrawimage.ProcessRawImage(file, self.spectrograph, self.det,
-                                                              self.proc_par)
-            processrawImage.process(self.process_steps, bias=bias, bpm=bpm)
-            # Instantiate the image stack
-            if image_arr is None:
-                shape = processrawImage.shape
-                image_arr = np.zeros(shape + (self.nfiles,))
-                self.binning = processrawImage.binning
-            # Check shape and binning consistency
-            if processrawImage.shape != shape:
-                msgs.error('All images to combine must have the same shape.')
-            if processrawImage.binning != self.binning:
-                msgs.error('All images to combine must have the same binning.')
-            # Store
-            image_arr[:,:,kk] = processrawImage.image
-
-        # Combine
-        self.image = image_arr[:,:,0] if self.nfiles == 1 \
-                else combine.comb_frames(image_arr,
-                                saturation=self.spectrograph.detector[self.det-1]['saturation'],
-                                method=self.proc_par['combine'], satpix=self.proc_par['satpix'],
-                                cosmics=self.proc_par['sigrej'], n_lohi=self.proc_par['n_lohi'],
-                                sig_lohi=self.proc_par['sig_lohi'],
-                                replace=self.proc_par['replace'])
-        # Return a copy image array for direct use
-        return self.image.copy()
-
+        buildImage = buildimage.BuildImage(self.spectrograph, self.det, self.proc_par, self.file_list)
+        self.pypeitImage = buildImage.run(self.process_steps, bias, bpm=bpm)
+        # Return
+        return self.pypeitImage
 
 
     def __repr__(self):