+++ CHANGES.rst
@@ -33,6 +33,13 @@
 - Fixed bug when bias was set to 'force' which was not bias subtracting
 - Implemented changes to vlt_xshooter_nir to now require darks taken between flats
 - Made flat fielding code a bit more robust against hot pixels at edge of orders
+- Added pypeit_chk_flat script to view flat images
+- Refactored image objects into RawImage, ProcessRawImage, PypeItImage, BuildImage
+- Moved load() and save() methods from MasterFrame to the individual calibration objects
+- Converted ArcImage and FlatImages into counts
+- Added code to allow for IVAR and RN2 image generation for calibs
+- Added several from_master_file() instantiation methods
+- Use coadd2d.weighted_combine() to stack calibration images
 - Major refactor of slit edge tracing (WIP)
 
 0.11.0 (22 Jun 2019)